requirements.txt
# Python依赖

# 核心依赖
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 训练工具
tqdm>=4.65.0
tensorboard>=2.14.0

# 配置和工具
pyyaml>=6.0
scipy>=1.10.0

# 评估指标
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.0

# 可选：LlamaFactory集成
# llamafactory-cli

# 可选：量化
# bitsandbytes>=0.41.0

# 可选：分布式训练
# deepspeed>=0.12.0
